requirements.txt
streamlit==1.29.0
langchain==0.0.306
numpy==1.26.2
Pillow==10.0.1
PyYAML==6.0.1